Changed 'long' by 'int' and 'unsigned long' by 'size_t' when necessary. long is 64 bit in GCC when compiling in 64 bits.

+++ include/Step/CallBack.h
@@ -25,8 +25,8 @@ namespace Step {
     {
     public:
         virtual ~CallBack() {}
-        virtual void setMaximum(unsigned long max) = 0;
-        virtual void setProgress(unsigned long progress) = 0;
+        virtual void setMaximum(size_t max) = 0;
+        virtual void setProgress(size_t progress) = 0;
     };
 }
